# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_murathanje_birthday_tracking_backend_internal_models.BirthdayResponse:
    description: Response model for birthday operations
    properties:
      birth_date:
        description: '@Description Birthday date (format: MM-DD)'
        example: 05-15
        type: string
      category_id:
        description: '@Description Category ID for grouping birthdays'
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      created_at:
        description: '@Description When the record was created'
        type: string
      id:
        description: '@Description Unique identifier for the birthday record'
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        description: '@Description Name of the person'
        example: John Doe
        type: string
      notes:
        description: '@Description Optional notes about the birthday'
        example: Best friend from college
        type: string
      updated_at:
        description: '@Description When the record was last updated'
        type: string
      user_id:
        description: '@Description User ID who owns this birthday record'
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
    type: object
  github_com_murathanje_birthday_tracking_backend_internal_models.Category:
    description: Category model for grouping birthdays
    properties:
      created_at:
        type: string
      icon:
        example: "\U0001F468‍\U0001F469‍\U0001F467‍\U0001F466"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Family
        type: string
    type: object
  github_com_murathanje_birthday_tracking_backend_internal_models.CreateBirthdayRequest:
    description: Request model for creating a birthday record
    properties:
      birth_date:
        description: '@Description Birthday date (format: MM-DD)'
        example: 05-15
        type: string
      category_id:
        description: '@Description Category ID for grouping birthdays'
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      name:
        description: '@Description Name of the person'
        example: John Doe
        type: string
      notes:
        description: '@Description Optional notes about the birthday'
        example: Best friend from college
        type: string
    required:
    - birth_date
    - category_id
    - name
    type: object
...
host: localhost:5050
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/murathanje/birthday_tracking_backend
  description: |-
    A birthday tracking service API in Go using Gin framework.
    Features:
    - User management with JWT authentication for user operations
    - API Key authentication for admin operations
    - Birthday tracking with categories
    - Category management with icons
    - Upcoming birthdays tracking

    Authentication:
    1. For Users:
    - Register a new account using /api/v1/register
    - Login with your credentials at /api/v1/login to get a JWT token
    - Use the token in the Authorization header for protected endpoints
    - Format: "Bearer <your_jwt_token>"
    2. For Admins:
    - Use API Key in the X-API-Key header for admin endpoints
    - The API Key should be set in your .env file
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Birthday Tracking API
  version: "1.0"
paths:
...
  /birthdays:
    get:
      description: Get all birthdays for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_murathanje_birthday_tracking_backend_internal_models.BirthdayResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get user's birthdays
      tags:
      - birthdays
    post:
      consumes:
      - application/json
      description: Create a new birthday record for the authenticated user
      parameters:
      - description: Birthday details
        in: body
        name: birthday
        required: true
        schema:
          $ref: '#/definitions/github_com_murathanje_birthday_tracking_backend_internal_models.CreateBirthdayRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_murathanje_birthday_tracking_backend_internal_models.BirthdayResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create a new birthday
      tags:
      - birthdays
  /birthdays/{id}:
    delete:
      description: Delete a birthday record (must belong to authenticated user)
      parameters:
      - description: Birthday ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete a birthday
      tags:
      - birthdays
    get:
      description: Get a birthday record by its ID (must belong to authenticated user)
      parameters:
      - description: Birthday ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_murathanje_birthday_tracking_backend_internal_models.BirthdayResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get a birthday by ID
      tags:
      - birthdays
    put:
      consumes:
      - application/json
      description: Update a birthday record (must belong to authenticated user)
      parameters:
      - description: Birthday ID
        in: path
        name: id
        required: true
        type: string
      - description: Birthday details
        in: body
        name: birthday
        required: true
        schema:
          $ref: '#/definitions/github_com_murathanje_birthday_tracking_backend_internal_models.CreateBirthdayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_murathanje_birthday_tracking_backend_internal_models.BirthdayResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update a birthday
      tags:
      - birthdays
  /categories:
    get:
      description: |-
        Get a list of all birthday categories with their icons
        Each category includes an ID, name, and an emoji icon
        Use this endpoint to populate category dropdowns or lists
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_murathanje_birthday_tracking_backend_internal_models.Category'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all categories
      tags:
      - categories
  /categories/{id}:
    get:
      description: Get a category by its ID
      parameters:
      - description: Category ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_murathanje_birthday_tracking_backend_internal_models.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a category by ID
      tags:
      - categories
  /categories/name/{name}:
    get:
      description: Get a category by its name (e.g., 'family', 'friends', etc.)
      parameters:
      - description: Category name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_murathanje_birthday_tracking_backend_internal_models.Category'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a category by name
      tags:
      - categories
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API Key required for admin operations. Set this in your .env file.
    in: header
    name: X-API-Key
    type: apiKey
  Bearer:
    description: Type "Bearer" followed by a space and JWT token. Required for user-specific
      operations.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...